refactor: Change the type of 'Point' to 'glm::dvec3'

+++ Includes/CubbyCity/Geometry/GeometryData.hpp
@@ -16,7 +16,7 @@
 
 namespace CubbyCity
 {
-using Point = glm::vec3;
+using Point = glm::dvec3;
 using Line = std::vector<Point>;
 using Polygon = std::vector<Line>;
